Added token authentication handling for the games API.
The token that is used to make calls to the IGDB API for the games now reside in the database instead of being an environment variable. the token has an expiry date and it'll get renewed automatically if it expired, ensuring the server doesn't break after the token validity duration passes.

I added a new middleware for this purpose and now all requests coming to the server will get the token in its header to use later and call the API.